+++ src/pages/Home/components/SearchItem.tsx
@@ -1,11 +1,11 @@
 import { Box, CircularProgress, Grid, Paper, Skeleton, Typography,makeStyles,styled, useTheme } from '@mui/material';
 import React, { useEffect, useMemo, useState } from 'react';
-import useSong from '../../hooks/song/useSong';
+import useSong from '../../../hooks/song/useSong';
 import { useNavigate } from 'react-router-dom'
-import Song from '../../models/song/song';
-import { SearchSongDataDTO } from '../../backend/dtos/dtosSong';
+import Song from '../../../models/song/song';
+import { SearchSongDataDTO } from '../../../backend/dtos/dtosSong';
 import {convertSheetToSections} from "@pepavlin/sheet-api";
-import useAuth from '../../hooks/auth/useAuth';
+import useAuth from '../../../hooks/auth/useAuth';
 
 const StyledContainer = styled(Box)(({theme})=>({
     backgroundColor: theme.palette.grey[100],
